I dont know why old controller test does not work, create new one to replace.
