+++ ApiControllerTest/BaseTest.cs
@@ -1,0 +1,11 @@
+﻿namespace ApiControllerTest
+{
+    public abstract class BaseTest
+    {
+        public BaseTest()
+        {
+            ConfigTest();
+        }
+        protected abstract void ConfigTest();
+    }
+}
