+++ ApiControllerTest/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="Castle.Core" version="4.1.1" targetFramework="net452" />
+  <package id="Moq" version="4.7.99" targetFramework="net452" />
+</packages>
