--- ControllerTest/BaseTest.cs
@@ -1,11 +1,0 @@
-﻿namespace ControllerTest
-{
-    public abstract class BaseTest
-    {
-        public BaseTest()
-        {
-            ConfigTest();
-        }
-        protected abstract void ConfigTest();
-    }
-}
